requirements.txt
fastapi==0.118.0
huggingface_hub==0.35.3
jose==1.0.0
matplotlib==3.10.6
openai==2.0.1
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
passlib==1.7.4
Pillow==11.3.0
pydantic==2.11.9
PyJWT==2.10.1
PyJWT==2.10.1
python_jose==3.5.0
Requests==2.32.5
scipy==1.16.2
SQLAlchemy==2.0.37
transformers==4.56.2
ultralytics==8.3.204

